requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
Flask-Migrate
Flask-JWT-Extended
python-dotenv
bcrypt
Flask-CORS
marshmallow
marshmallow-sqlalchemy
